Add support for file count threshold and commit hashes threshold
Whenever a user calls the command, before this commit - Context Pilot will return all the relevant files and run through the git commit hashes with threshold of 5.

After this commit, the users will have option to set the thresholds themselves in $HOME_DIR/context_pilot.json, if not present - we'll use the defaults (defined in `config.rs` file). A small TODO after this is to make sure that users are pointed for the invalid JSON file format and also they should be able to skip one of the variables that need to be defined.

Ideally, I'm planning to switch to Lua for the config file instead of JSON, for now this is fine.