A failed attempt on supporting partial indexing.
Partial Indexing: Let ContextPilot attempt indexing for the whole workspace only once. Store the last commit in the metadata, and when the next time query is done and there are commits that have not been indexed - index "just for them" and update the DB + metadata (optionally).

Issue here is that the diff logic needs to rebuild the map then from the particular file - and while it's possible, I'm assuming there'll be some bugs.

I'll pick this again though, as it's a lovely feature to have.